www/Flask/nicoapi/requirements.txt
google-cloud-firestore
google-cloud-storage
urllib3
certifi
requests
